# cmd: make coreSupportsReExec faster