--- a/cmd/cmd_linux_test.go
+++ b/cmd/cmd_linux_test.go
@@ -61,13 +61,11 @@ Nullam euismod neque ut turpis molestie, eu fringilla elit volutpat.
 Phasellus maximus, urna eget porta congue, diam enim volutpat diam, nec ultrices lorem risus ac metus.
 Vivamus convallis eros non nunc pretium bibendum.
 Maecenas consectetur metus metus.
-Orci varius natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus.
 Morbi scelerisque urna at arcu tristique feugiat.
 Vestibulum condimentum odio sed tortor vulputate, eget hendrerit mi consequat.
 Integer egestas finibus augue, ac scelerisque ex pretium aliquam.
 Aliquam erat volutpat.
 Suspendisse a nulla ultrices, porttitor tellus ut, bibendum diam.
-Orci varius natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus.
 In nibh dui, tempus eget vestibulum in, euismod in ex.
 In tempus felis lectus.