coreSupportsReExec is rather critical performance-wise, as it's used
for every invocation of snap, including every binary in /snap/bin.

So I thought I might try to speed it up a bit. This ended up with some
work in `strutil` which got split into #6074, but there was still room
for improvement here. Even with that on master, this change helps:

```
$ benchcmp /tmp/master.txt /tmp/this.txt
benchmark                    old ns/op     new ns/op     delta
BenchmarkCSRE_real-8         18511         4256          -77.01%

benchmark                    old allocs     new allocs     delta
BenchmarkCSRE_real-8         50             12             -76.00%

benchmark                    old bytes     new bytes     delta
BenchmarkCSRE_real-8         44041         1176          -97.33%
```

HTH, HAND.