@zyga 2018-10-30
As a quick comment I would feel more comfortable reviewing this as separate patches.