@chipaca 2018-10-30
@mvo5 @bboozzoo 811eda2dc4f91f2806ae5a363dcaac36776c977b would make arch-ish 1-2-3 versions work as well, I can add it to this if you think it's a good idea.
Or I can split this into a 'revamp versionn.go' and the cmd changes, to keep @zyga happy.